# perf(synthetic-shadow): optimize patchedAddEventListener