--- a/packages/@lwc/synthetic-shadow/src/polyfills/event-target/polyfill.ts
+++ b/packages/@lwc/synthetic-shadow/src/polyfills/event-target/polyfill.ts
@@ -19,22 +19,26 @@ import { getEventListenerWrapper } from '../../shared/event-target';
 
 function patchedAddEventListener(
     this: EventTarget,
-    _type: string,
-    _listener: EventListenerOrEventListenerObject,
-    _optionsOrCapture?: boolean | AddEventListenerOptions
+    type: string,
+    listener: EventListenerOrEventListenerObject,
+    optionsOrCapture?: boolean | AddEventListenerOptions
 ) {
     if (isHostElement(this)) {
         // Typescript does not like it when you treat the `arguments` object as an array
         // @ts-ignore type-mismatch
         return addCustomElementEventListener.apply(this, arguments);
     }
-    const args = ArraySlice.call(arguments);
-    if (args.length > 1) {
-        args[1] = getEventListenerWrapper(args[1]);
+    if (arguments.length < 2) {
+        // The browser throws a TypeError for insufficient arguments, so we should too.
+        // https://googlechrome.github.io/samples/event-listeners-mandatory-arguments/
+        throw new TypeError(
+            `Failed to execute 'addEventListener' on 'EventTarget': 2 arguments required, but only ${arguments.length} present.`
+        );
     }
-    // Ignore types because we're passing through to native method
-    // @ts-ignore type-mismatch
-    return nativeAddEventListener.apply(this, args);
+
+    const wrappedListener = getEventListenerWrapper(listener) as EventListenerOrEventListenerObject;
+    // The third argument is optional, so passing in `undefined` for `optionsOrCapture` gives capture=false
+    return nativeAddEventListener.call(this, type, wrappedListener, optionsOrCapture);
 }
 
 function patchedRemoveEventListener(