--- a/packages/@lwc/synthetic-shadow/src/polyfills/event-target/polyfill.ts
+++ b/packages/@lwc/synthetic-shadow/src/polyfills/event-target/polyfill.ts
@@ -29,11 +29,15 @@ function patchedAddEventListener(
         return addCustomElementEventListener.apply(this, arguments);
     }
     if (arguments.length < 2) {
-        // The browser throws a TypeError for insufficient arguments, so we should too.
+        // Slow path, unlikely to be called frequently. We expect modern browsers to throw:
         // https://googlechrome.github.io/samples/event-listeners-mandatory-arguments/
-        throw new TypeError(
-            `Failed to execute 'addEventListener' on 'EventTarget': 2 arguments required, but only ${arguments.length} present.`
-        );
+        const args = ArraySlice.call(arguments);
+        if (args.length > 1) {
+            args[1] = getEventListenerWrapper(args[1]);
+        }
+        // Ignore types because we're passing through to native method
+        // @ts-ignore type-mismatch
+        return nativeAddEventListener.apply(this, args);
     }
 
     const wrappedListener = getEventListenerWrapper(listener) as EventListenerOrEventListenerObject;